requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
SQLAlchemy
tqdm
pydantic
